Refactor url.
